refactor(content): update desc
