+++ src/styles/Hero.css
@@ -59,7 +59,7 @@
     padding-right: 1.5rem;
     background: rgba(0, 0, 0, 0.1);
     text-shadow:  8px 6px 10px black;
-    backdrop-filter: blur(8px);
+    backdrop-filter: blur(3px);
     border-radius: 20px;
     color: white;
 }
@@ -77,7 +77,7 @@
 }
 
 .hero-name {
-    font-size: 2.75rem;
+    font-size: 2.85rem;
     font-weight: 700;
     margin-bottom: 1rem;
     margin-top: 2rem;
@@ -95,7 +95,7 @@
 }
 
 .hero-role {
-    font-size: 2rem;
+    font-size: 1.7rem;
     font-weight: 500;
     color: #fdfdfd;
     margin-bottom: 1.5rem;
